+++ .github/workflows/ci.yml
@@ -39,7 +39,7 @@ jobs:
               - 'libs/checkpoint/**'
               - 'libs/checkpoint-sqlite/**'
               - 'libs/checkpoint-postgres/**'
-              - 'libs/checkpoint-validation/**'
+              - 'libs/checkpoint-conformance/**'
               - 'libs/prebuilt/**'
             deps:
               - '**/pyproject.toml'
@@ -58,7 +58,7 @@ jobs:
             "libs/checkpoint",
             "libs/checkpoint-sqlite",
             "libs/checkpoint-postgres",
-            "libs/checkpoint-validation",
+            "libs/checkpoint-conformance",
             "libs/prebuilt",
           ]
     if: needs.changes.outputs.python == 'true' || needs.changes.outputs.deps == 'true'
@@ -78,7 +78,7 @@ jobs:
             "libs/checkpoint",
             "libs/checkpoint-sqlite",
             "libs/checkpoint-postgres",
-            "libs/checkpoint-validation",
+            "libs/checkpoint-conformance",
             "libs/prebuilt",
             "libs/sdk-py",
           ]
