+++ libs/checkpoint-conformance/langgraph/checkpoint/conformance/spec/__init__.py
@@ -1,0 +1,27 @@
+"""Test spec modules for each checkpointer capability."""
+
+from langgraph.checkpoint.conformance.spec.test_copy_thread import (
+    run_copy_thread_tests,
+)
+from langgraph.checkpoint.conformance.spec.test_delete_for_runs import (
+    run_delete_for_runs_tests,
+)
+from langgraph.checkpoint.conformance.spec.test_delete_thread import (
+    run_delete_thread_tests,
+)
+from langgraph.checkpoint.conformance.spec.test_get_tuple import run_get_tuple_tests
+from langgraph.checkpoint.conformance.spec.test_list import run_list_tests
+from langgraph.checkpoint.conformance.spec.test_prune import run_prune_tests
+from langgraph.checkpoint.conformance.spec.test_put import run_put_tests
+from langgraph.checkpoint.conformance.spec.test_put_writes import run_put_writes_tests
+
+__all__ = [
+    "run_put_tests",
+    "run_put_writes_tests",
+    "run_get_tuple_tests",
+    "run_list_tests",
+    "run_delete_thread_tests",
+    "run_delete_for_runs_tests",
+    "run_copy_thread_tests",
+    "run_prune_tests",
+]
