+++ libs/checkpoint-conformance/langgraph/checkpoint/conformance/spec/test_get_tuple.py
@@ -8,7 +8,7 @@ from uuid import uuid4
 
 from langgraph.checkpoint.base import BaseCheckpointSaver
 
-from langgraph.checkpoint.validation.test_utils import (
+from langgraph.checkpoint.conformance.test_utils import (
     generate_checkpoint,
     generate_config,
     generate_metadata,
